feature : User - 회원가입, 로그인 동시 진행 - 동명이인일시 방법 필요

+++ src/main/java/com/future/yearend/service/MemoService.java
@@ -1,13 +1,15 @@
 package com.future.yearend.service;
 
+import com.future.yearend.common.JwtUtil;
 import com.future.yearend.entity.Memo;
 import com.future.yearend.dto.MemoRequestDto;
 import com.future.yearend.dto.MemoResponseDto;
+import com.future.yearend.entity.User;
 import com.future.yearend.repository.MemoRepository;
+import com.future.yearend.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -16,10 +18,15 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class MemoService {
     private final MemoRepository memoRepository;
+    private final UserRepository userRepository;
+    private final JwtUtil jwtUtil;
 
-    public MemoResponseDto createMemo(MemoRequestDto memoRequestDto) {
-        Memo memo = memoRepository.save(new Memo(memoRequestDto));
-        return new MemoResponseDto(memo);
+    public MemoResponseDto createMemo(MemoRequestDto memoRequestDto, String username) {
+        User user = findUser(username);
+        Memo memo = new Memo(memoRequestDto, user);
+
+        Memo saveMemo = memoRepository.save(memo);
+        return new MemoResponseDto(saveMemo, user);
     }
 
     public List<MemoResponseDto> getAllMemo() {
@@ -32,14 +39,28 @@ public class MemoService {
     }
 
 
-    public ResponseEntity<String> updateMemo(Long id, MemoRequestDto memoRequestDto) {
+    public ResponseEntity<String> updateMemo(Long id, MemoRequestDto memoRequestDto, String username) {
         Memo memo = findMemo(id);
-        memo.update(memoRequestDto);
+        User user = findUser(username);
+
+        if (!memo.getUser().getUsername().equals(username)) {
+            throw new IllegalArgumentException("작성자가 다릅니다.");
+        }
+
+        // 메모 업데이트
+        memo.update(memoRequestDto, user);
+        memoRepository.save(memo);
         return ResponseEntity.ok("수정 성공!");
     }
 
-    public ResponseEntity<String> deleteMemo(Long id) {
+    public ResponseEntity<String> deleteMemo(Long id, String username) {
         Memo memo = findMemo(id);
+        User user = findUser(username);
+
+        if (!memo.getUser().getUsername().equals(username)) {
+            throw new IllegalArgumentException("작성자가 다릅니다.");
+        }
+
         memoRepository.delete(memo);
         return ResponseEntity.ok("삭제 성공!");
     }
@@ -49,6 +70,11 @@ public class MemoService {
     private Memo findMemo(Long id) {
         return memoRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("해당 메모는 존재하지 않습니다.")
+        );
+    }
+    private User findUser(String username) {
+        return userRepository.findByUsername(username).orElseThrow(
+                ()-> new IllegalArgumentException("해당 작성자는 존재하지 않습니다.")
         );
     }
 }
